Fixed different bugs
- Now you can reset when parameter torqueBalancingSequence is set to
torqueBalancing.
- Switch dequeue<int> to dequeue<double> for compatibility with
WBI-Toolbox.

+++ src/modules/walkPlayer/include/walkPlayer/actionClass.h
@@ -35,7 +35,7 @@ struct actionStructForTorqueBalancing
 {
     std::deque< std::deque<double> >  com_traj;
     std::deque< std::deque<double> >  postural_traj;
-    std::deque< std::deque<int> >     constraints;
+    std::deque< std::deque<double> >     constraints;
     
 public:
     actionStructForTorqueBalancing();
